--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,10 @@
+module go-metrics
+
+go 1.13
+
+require (
+	github.com/ajstarks/svgo v0.0.0-20190826172357-de52242f3d65 // indirect
+	github.com/rcrowley/go-metrics v0.0.0-20190826022208-cac0b30c2563
+	github.com/stathat/go v1.0.0
+	golang.org/x/tools v0.0.0-20191107235519-f7ea15e60b12 // indirect
+)