--- a/registry.go
+++ b/registry.go
@@ -22,7 +22,6 @@ func (err DuplicateMetric) Error() string {
 // This is an interface so as to encourage other structs to implement
 // the Registry API as appropriate.
 type Registry interface {
-
 	// Call the given function for each registered metric.
 	Each(func(string, interface{}))
 
@@ -53,27 +52,27 @@ type Registry interface {
 // The standard implementation of a Registry is a mutex-protected map
 // of names to metrics.
 type StandardRegistry struct {
-	metrics map[string]interface{}
-	mutex   sync.Mutex
+	metrics sync.Map
 }
 
 // Create a new registry.
 func NewRegistry() Registry {
-	return &StandardRegistry{metrics: make(map[string]interface{})}
+	return &StandardRegistry{metrics: sync.Map{}}
 }
 
 // Call the given function for each registered metric.
 func (r *StandardRegistry) Each(f func(string, interface{})) {
-	for name, i := range r.registered() {
-		f(name, i)
-	}
+	registered := r.registered()
+	registered.Range(func(name, i interface{}) bool {
+		f(name.(string), i)
+		return true
+	})
 }
 
 // Get the metric by the given name or nil if none is registered.
 func (r *StandardRegistry) Get(name string) interface{} {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-	return r.metrics[name]
+	val, _ := r.metrics.Load(name)
+	return val
 }
 
 // Gets an existing metric or creates and registers a new one. Threadsafe
@@ -81,9 +80,7 @@ func (r *StandardRegistry) Get(name string) interface{} {
 // The interface can be the metric to register if not found in registry,
 // or a function returning the metric for lazy instantiation.
 func (r *StandardRegistry) GetOrRegister(name string, i interface{}) interface{} {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-	if metric, ok := r.metrics[name]; ok {
+	if metric, ok := r.metrics.Load(name); ok {
 		return metric
 	}
 	if v := reflect.ValueOf(i); v.Kind() == reflect.Func {
@@ -96,20 +93,17 @@ func (r *StandardRegistry) GetOrRegister(name string, i interface{}) interface{}
 // Register the given metric under the given name.  Returns a DuplicateMetric
 // if a metric by the given name is already registered.
 func (r *StandardRegistry) Register(name string, i interface{}) error {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
 	return r.register(name, i)
 }
 
 // Run all registered healthchecks.
 func (r *StandardRegistry) RunHealthchecks() {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-	for _, i := range r.metrics {
+	r.metrics.Range(func(_, i interface{}) bool {
 		if h, ok := i.(Healthcheck); ok {
 			h.Check()
 		}
-	}
+		return true
+	})
 }
 
 // GetAll metrics in the Registry
@@ -175,45 +169,41 @@ func (r *StandardRegistry) GetAll() map[string]map[string]interface{} {
 
 // Unregister the metric with the given name.
 func (r *StandardRegistry) Unregister(name string) {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
 	r.stop(name)
-	delete(r.metrics, name)
+	r.metrics.Delete(name)
 }
 
 // Unregister all metrics.  (Mostly for testing.)
 func (r *StandardRegistry) UnregisterAll() {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-	for name, _ := range r.metrics {
-		r.stop(name)
-		delete(r.metrics, name)
-	}
+	r.metrics.Range(func(name, _ interface{}) bool {
+		r.stop(name.(string))
+		r.metrics.Delete(name)
+		return true
+	})
 }
 
 func (r *StandardRegistry) register(name string, i interface{}) error {
-	if _, ok := r.metrics[name]; ok {
+	if _, ok := r.metrics.Load(name); ok {
 		return DuplicateMetric(name)
 	}
 	switch i.(type) {
 	case Counter, Gauge, GaugeFloat64, Healthcheck, Histogram, Meter, Timer:
-		r.metrics[name] = i
+		r.metrics.Store(name, i)
 	}
 	return nil
 }
 
-func (r *StandardRegistry) registered() map[string]interface{} {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-	metrics := make(map[string]interface{}, len(r.metrics))
-	for name, i := range r.metrics {
-		metrics[name] = i
-	}
+func (r *StandardRegistry) registered() sync.Map {
+	metrics := sync.Map{}
+	r.metrics.Range(func(name, i interface{}) bool {
+		metrics.Store(name, i)
+		return true
+	})
 	return metrics
 }
 
 func (r *StandardRegistry) stop(name string) {
-	if i, ok := r.metrics[name]; ok {
+	if i, ok := r.metrics.Load(name); ok {
 		if s, ok := i.(Stoppable); ok {
 			s.Stop()
 		}