Удалены мутексы

До:
```
BenchmarkGetOrRegisterCounter-8              	 6892653	       175 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetOrRegisterCounter_RandomName-8   	  559737	      2184 ns/op	     481 B/op	      12 allocs/op
```

После:
```
BenchmarkGetOrRegisterCounter-8              	92660134	        13.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetOrRegisterCounter_RandomName-8   	  454476	      3307 ns/op	     592 B/op	      16 allocs/op
```